Struct recycler::TrashRecycler

pub struct TrashRecycler<Item> {
    // some fields omitted
}

A "recycler" that doesn't recycle anything, instead just dropping anything it is given. This is particularly useful for primitive types such as i32 that do not have Drop implementations.